Change how the calculator handle the numbers
Beforehand, the calculator was getting the first
number, the operator and the second number and
throwing all these items into an array.

After pass all these items to an array, there were
functions to try to separate all these items again
to the corresponding part. So, numbers located before
the operator were the first number, those after the
operator were the second number, etc.

As you can see, it wasn't very smart. Basically
we were putting pieces together to separate them again.
Also, it was creating a lot of bugs and unexpected
problems.

The fix was to give the first number it's own list,
repeat the same for the second number, and check
the operator to make an operation.

For sure, there are some tricks involved, mainly to
check the value of an operator and how to track the
result, but nothing extravagant.